# roachpb: Use make and copy pattern for high profile byte slice copies