@tamird 2016-03-08
LGTM. Seems like this deserves a bug report upstream, too.