@nvanbenschoten 2016-03-08
@tamird It's not necessarily a bug, it's just that the one-liner adds a few extra instructions and uses `runtime.growslice_n` instead of `runtime.makeslice`.