@tamird 2016-03-08
Why isn't that a bug? Is there any difference in behaviour? Seems like bad codegen to me.