@nvanbenschoten 2016-03-08
Filed it upstream https://github.com/golang/go/issues/14718. I'm guessing it will expressed as expected behavior, but I'm curious what the Go team says nevertheless.